Move prisma helpers
