+++ packages/api/src/controllers/attributes-controller.ts
@@ -1,6 +1,6 @@
 import { ApiError } from '@/lib/api-error.js';
 import { apiResponse } from '@/lib/api-response/index.js';
-import { prisma } from '@/lib/prisma/instance.js';
+import { prisma } from '@/lib/prisma-instance.js';
 import { idParamsSchema } from '@/lib/schemas.js';
 import { createAttributeSchema, updateAttributeSchema } from '@optimism-making-impact/schemas';
 import { NextFunction, Request, Response } from 'express';
